requirements.txt
pandas>=2.0
numpy>=1.24
plotly>=5.18
streamlit>=1.32
requests>=2.31
python-dotenv>=1.0
